shadow-dom/Element-interface-attachShadow-custom-element.html

<!DOCTYPE html>
<title>Shadow DOM: Attaching a ShadowRoot for custom elements</title>
<meta name="author" title="Hayato Ito" href="mailto:hayato@chromium.org">
<link rel="help" href="https://dom.spec.whatwg.org/#dom-element-attachshadow">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
class MyAutonomousCustomElement extends HTMLElement {
}

customElements.define('my-custom', MyAutonomousCustomElement);

test(() => {
  assert_true(document.createElement('my-custom').attachShadow({mode: "open"}) instanceof ShadowRoot);
}, 'Element.attachShadow must create an instance of ShadowRoot for autonomous custom elements');

test(() => {
  assert_true(document.createElement('undefined-custom').attachShadow({mode: 'open'}) instanceof ShadowRoot);
}, 'Element.attachShadow must create an instance of ShadowRoot for undefined autonomous custom elements');

test(() => {
  class ShadowDisabledElement extends HTMLElement {
    static disabledFeatures = ['shadow'];
  }

  // No definition. attachShadow() should succeed.
  let element = document.createElement('shadow-disabled-element');
  element.attachShadow({mode: 'closed'});

  // No definition and it's already a host.
  assert_throws('NotSupportedError', () => {
    element.attachShadow({mode: 'closed'});
  });

  // The element has a definition, and it's already a host.
  customElements.define('shadow-disabled-element', ShadowDisabledElement);
  assert_throws('NotSupportedError', () => {
    element.attachShadow({mode: 'closed'});
  });
}, 'Element.attachShadow for a custom element with disabledFeatures=["shadow"] should throw a NotSupportedError');

test(() => {
  class CapitalShadowDisabledElement extends HTMLElement {
    static disabledFeatures = ['SHADOW'];
  }

  customElements.define('capital-shadow-disabled-element', CapitalShadowDisabledElement);

  // Test fails if this throws
  document.createElement('capital-shadow-disabled-element').attachShadow({mode: 'open'});
}, 'Element.attachShadow for a custom element with disabledFeatures=["SHADOW"] should not throw a NotSupportedError');

class MyCustomizedBuiltinElement extends HTMLInputElement {
}

customElements.define('my-input', MyCustomizedBuiltinElement, { extends: 'input' });

test(() => {
  assert_throws({'name': 'NotSupportedError'}, () => {
    document.createElement('input', {is: 'my-input'}).attachShadow({mode: "open"});
  });
}, 'Element.attachShadow must throw a NotSupportedError for customized built-in elements');
</script>
</body>
</html>
